# common-library/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: E-Learning Platform API
  description: API for managing users, courses, payments, recommendations, and more.
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging.example.com/v1
    description: Staging server
paths:
...
  /courses:
    get:
      summary: List all courses
      responses:
        '200':
          description: A list of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
    post:
      summary: Create a new course
      requestBody:
        description: Course to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
        required: true
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
  /courses/{courseId}:
    get:
      summary: Get a course by ID
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          description: Course not found
    put:
      summary: Update a course
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      requestBody:
        description: Updated course information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
        required: true
      responses:
        '200':
          description: Course updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          description: Course not found
    delete:
      summary: Delete a course
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Course deleted
        '404':
          description: Course not found

  # Payments
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        created_at:
          type: string
          format: date-time
    CourseInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
...